CORE-650: Add SFTP account status UI
In the Org Data Autoloader Field Mappings page, add the UI
(that has static text now) to show the host and port of the
SFTP site followed by the account status, which is either
Active or Not Created.

Update top-level HTML structure to be a list group.

Remove component level styling in favor of Bootstrap theme.

Use the new Font Awesome.

Fix linting errors.